model/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pymupdf 
langchain 
langchain-community 
chromadb 
sentence-transformers 
accelerate
google-generativeai